python/pyproject.toml

[build-system]
requires = ["setuptools>=67"]
build-backend = "setuptools.build_meta"

[project]
name = "antenna-spec-tools"
version = "0.1.0"
description = "Reference tools for Antenna Distributed Agent Social Protocol spec"
requires-python = ">=3.10"
dependencies = [
  "jsonschema>=4.0",
  "pycryptodome>=3.0",
]

[tool.setuptools.packages.find]
where = ["python"]
